+++ x/evm/types/params.go
@@ -37,7 +37,7 @@ var (
 	}
 	// DefaultExtraEIPs defines the default extra EIPs to be included
 	// On v15, EIP 3855 was enabled
-	DefaultExtraEIPs   = []int64{3855}
+	DefaultExtraEIPs   = []string{"ethereum_3855"}
 	DefaultEVMChannels = []string{
 		"channel-10", // Injective
 		"channel-31", // Cronos
@@ -62,7 +62,7 @@ func NewParams(
 	evmDenom string,
 	allowUnprotectedTxs bool,
 	config ChainConfig,
-	extraEIPs []int64,
+	extraEIPs []string,
 	activeStaticPrecompiles,
 	evmChannels []string,
 	accessControl AccessControl,
@@ -138,12 +138,10 @@ func (p Params) Validate() error {
 	return validateChannels(p.EVMChannels)
 }
 
-// EIPs returns the ExtraEIPS as a int slice
-func (p Params) EIPs() []int {
-	eips := make([]int, len(p.ExtraEIPs))
-	for i, eip := range p.ExtraEIPs {
-		eips[i] = int(eip)
-	}
+// EIPs returns the ExtraEIPS as a slice.
+func (p Params) EIPs() []string {
+	eips := make([]string, len(p.ExtraEIPs))
+	copy(eips, p.ExtraEIPs)
 	return eips
 }
 
@@ -232,22 +230,27 @@ func validateBool(i interface{}) error {
 }
 
 func validateEIPs(i interface{}) error {
-	eips, ok := i.([]int64)
+	eips, ok := i.([]string)
 	if !ok {
 		return fmt.Errorf("invalid EIP slice type: %T", i)
 	}
 
-	uniqueEIPs := make(map[int64]struct{})
+	uniqueEIPs := make(map[string]struct{})
 
 	for _, eip := range eips {
-		if !vm.ExistsEipActivator(int(eip)) {
-			return fmt.Errorf("EIP %d is not activateable, valid EIPs are: %s", eip, vm.ActivateableEips())
+		if !vm.ExistsEipActivator(eip) {
+			return fmt.Errorf("EIP %s is not activateable, valid EIPs are: %s", eip, vm.ActivateableEips())
+		}
+
+		if err := vm.ValidateEIPName(eip); err != nil {
+			return fmt.Errorf("EIP %s name is not valid", eip)
 		}
 
 		if _, ok := uniqueEIPs[eip]; ok {
-			return fmt.Errorf("found duplicate EIP: %d", eip)
+			return fmt.Errorf("found duplicate EIP: %s", eip)
 		}
 		uniqueEIPs[eip] = struct{}{}
+
 	}
 
 	return nil